+++ web/test/test-lib.js
@@ -291,7 +291,8 @@ Side.prototype = {
   },
 
   setSampleRadius: function(radius) {
-    var slider = this.side.element(by.id('sample-radius-slider'));
+    this.side.$('#setting-sample-radius').click();
+    var slider = $('#sample-radius-slider');
     slider.getAttribute('value').then(function(value) {
       var diff = radius - value;
       while (diff > 0) {
@@ -340,6 +341,12 @@ Side.prototype = {
 
   startSqlSaving: function() {
     this.side.element(by.id('save-results-opener')).click();
+  },
+
+  clickSqlSort(colId) {
+    var res = this.side.$('#sql-result');
+    var header = res.$$('thead tr th').get(colId);
+    header.click();
   },
 
   executeSqlSaving: function() {
@@ -958,7 +965,7 @@ testLib = {
           }});
       return defer.promise;
     }
-    browser.controlFlow().execute(sendRequest);
+    return browser.controlFlow().execute(sendRequest);
   },
 
   navigateToProject: function(name) {
@@ -1011,7 +1018,7 @@ testLib = {
               e.element(by.cssContainingText('option', optionLabelPattern)).click();
             }
           } else if (kind === 'choice') {
-            e.element(by.cssContainingText('option', value)).click();
+            e.$('option[label="' + value +'"]').click();
           } else {
             e.sendKeys(testLib.selectAllKey + value);
           }
@@ -1175,6 +1182,15 @@ testLib = {
 
   showSelector: function() {
     $('#show-selector-button').click();
+  },
+
+  confirmSweetAlert: function(expectedMessage) {
+    // SweetAlert is not an Angular library. We need to wait until it pops in and out.
+    var EC = protractor.ExpectedConditions;
+    testLib.wait(EC.visibilityOf($('.sweet-alert.showSweetAlert.visible')));
+    expect($('.sweet-alert h2').getText()).toBe(expectedMessage);
+    $('.sweet-alert button.confirm').click();
+    testLib.wait(EC.stalenessOf($('.sweet-alert.showSweetAlert')));
   },
 };