+++ CHANGELOG.md
@@ -6,11 +6,16 @@ Please add changes to "master", preferably ordered by their significance. (Most 
 
 ### master
 
- - Tab-completion for LynxKite operations works in Python command-line 
+ - Visualization settings UI now matches the new attribute UI.
+ - When you try to overwrite a table or view, you now get a confirmation prompt instead of an error.
+ - Tab-completion for LynxKite operations works in Python command-line
    and in Jupyter notebooks.
+ - You can now filter for position (Double, Double) attributes.
+ - Added _Segment by geographical proximity_ operation to segment vertices using Shapefiles and
+   coordinate vertex attributes. Can be used e.g. to segment by geo regions or points of interests.
  - New _"Sample graph by random walks"_ operation added. It can be used to make small smart samples
    that preserves the structure of the network to speed up model creation.
-   
+
 ### 1.12.1
 
  - Admins can backup LynxKite project data from UI.
@@ -36,6 +41,7 @@ Please add changes to "master", preferably ordered by their significance. (Most 
  - Spark Monitor doesn't send alarm in case of long-running tasks.
 
 Ecosystem:
+
  - Replace `run_task.sh` and `show_target.sh` with a new unified `tasks.sh`.
  - Renamed from "Lynx Data Automation Framework" to "Lynx Enterprise".